Velha e Hanoi modularizados
